requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aioredis==1.3.1
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
autoflake==2.2.0
bcrypt==4.0.1
beanie==1.20.0
bson==0.5.10
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
factory-boy==3.2.1
Faker==18.10.1
fastapi==0.95.0
filelock==3.12.2
flake8==6.0.0
frozenlist==1.3.3
h11==0.14.0
hiredis==2.2.3
httpcore==0.17.2
httpx==0.24.1
identify==2.5.24
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
lazy-model==0.0.5
load-dotenv==0.1.0
MarkupSafe==2.1.3
mccabe==0.7.0
mock==5.0.2
mongomock==4.1.2
motor==3.1.2
motor-stubs==1.7.1
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pika==1.3.2
pika-stubs==0.1.3
platformdirs==3.8.0
pluggy==1.0.0
pre-commit==3.3.3
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pyflakes==3.0.1
PyJWT==2.7.0
pymongo==4.3.3
PySocks==1.7.1
pytest==7.3.2
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
selenium==4.10.0
sentinels==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.26.1
toml==0.10.2
trio==0.22.1
trio-websocket==0.10.3
twilio==8.4.0
twilio-stubs==0.2.0
types-docutils==0.20.0.1
types-mock==5.0.0.7
types-openpyxl==3.1.0.13
types-passlib==1.7.7.12
types-Pygments==2.15.0.1
types-setuptools==68.0.0.0
typing_extensions==4.6.3
urllib3==2.0.3
virtualenv==20.23.1
wsproto==1.2.0
yarl==1.9.2
